// packages/core/src/components/fiatWallet/CreateFiatWallet/type.ts
export enum CreateFiatWalletSteps {
  WHATS_NEW = 'whatsNew',
  INTRO = 'intro',
  TELL_US_ABOUT_YOU = 'tellUsAboutYou',
  CHOOSE_CURRENCY = 'chooseCurrency',
  ACTIVATION_SUCCESS = 'success',
  CHOOSE_DOCUMENT = 'chooseDocument',
  UPLOAD = 'upload',
  HANDLE_ID_REVIEW_ERROR = 'handleIdReviewError',
  DOCUMENT_UNDER_REVIEW = 'documentUnderReview',
  REVIEW_INFO_BEFORE_ADDING_ID = 'reviewInfoBeforeAddingId',
}